refactor: move readline
